// src/types/tamagochai.ts
export type AvatarExpression =
  | 'neutral'
  | 'happy'
  | 'sad'
  | 'angry'
  | 'scared'
  | 'loving';